add server manager
