+++ cdc/owner/owner_test.go
@@ -17,7 +17,6 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"math"
 	"testing"
 	"time"
 
@@ -80,8 +79,6 @@ func newOwner4Test(
 ) Owner {
 	m := upstream.NewManager4Test(pdClient)
 	o := NewOwner(m, config.NewDefaultSchedulerConfig()).(*ownerImpl)
-	// Most tests do not need to test bootstrap.
-	o.bootstrapped = true
 	o.newChangefeed = func(
 		id model.ChangeFeedID,
 		state *orchestrator.ChangefeedReactorState,
@@ -271,8 +268,6 @@ func TestStopChangefeed(t *testing.T) {
 func TestFixChangefeedState(t *testing.T) {
 	ctx := cdcContext.NewBackendContext4Test(false)
 	owner, state, tester := createOwner4Test(ctx, t)
-	// We need to do bootstrap.
-	owner.bootstrapped = false
 	changefeedID := model.DefaultChangeFeedID("test-changefeed")
 	// Mismatched state and admin job.
 	changefeedInfo := &model.ChangeFeedInfo{
@@ -293,7 +288,6 @@ func TestFixChangefeedState(t *testing.T) {
 	_, err = owner.Tick(ctx, state)
 	tester.MustApplyPatches()
 	require.Nil(t, err)
-	require.True(t, owner.bootstrapped)
 	require.NotContains(t, owner.changefeeds, changefeedID)
 	// Start tick normally.
 	_, err = owner.Tick(ctx, state)
@@ -307,8 +301,6 @@ func TestFixChangefeedState(t *testing.T) {
 func TestFixChangefeedSinkProtocol(t *testing.T) {
 	ctx := cdcContext.NewBackendContext4Test(false)
 	owner, state, tester := createOwner4Test(ctx, t)
-	// We need to do bootstrap.
-	owner.bootstrapped = false
 	changefeedID := model.DefaultChangeFeedID("test-changefeed")
 	// Unknown protocol.
 	changefeedInfo := &model.ChangeFeedInfo{
@@ -333,7 +325,6 @@ func TestFixChangefeedSinkProtocol(t *testing.T) {
 	_, err = owner.Tick(ctx, state)
 	tester.MustApplyPatches()
 	require.Nil(t, err)
-	require.True(t, owner.bootstrapped)
 	require.NotContains(t, owner.changefeeds, changefeedID)
 
 	// Start tick normally.
@@ -438,113 +429,6 @@ func TestAdminJob(t *testing.T) {
 		},
 	})
 	require.Len(t, owner.takeOwnerJobs(), 0)
-}
-
-func TestUpdateGCSafePoint(t *testing.T) {
-	mockPDClient := &gc.MockPDClient{}
-	m := upstream.NewManager4Test(mockPDClient)
-	o := NewOwner(m, config.NewDefaultSchedulerConfig()).(*ownerImpl)
-	ctx := cdcContext.NewBackendContext4Test(true)
-	ctx, cancel := cdcContext.WithCancel(ctx)
-	defer cancel()
-	state := orchestrator.NewGlobalState(etcd.DefaultCDCClusterID)
-	tester := orchestrator.NewReactorStateTester(t, state, nil)
-
-	// no changefeed, the gc safe point should be max uint64
-	mockPDClient.UpdateServiceGCSafePointFunc = func(
-		ctx context.Context, serviceID string, ttl int64, safePoint uint64,
-	) (uint64, error) {
-		// Owner will do a snapshot read at (checkpointTs - 1) from TiKV,
-		// set GC safepoint to (checkpointTs - 1)
-		require.Equal(t, safePoint, uint64(math.MaxUint64-1))
-		return 0, nil
-	}
-	err := o.updateGCSafepoint(ctx, state)
-	require.Nil(t, err)
-
-	// add a failed changefeed, it must not trigger update GC safepoint.
-	mockPDClient.UpdateServiceGCSafePointFunc = func(
-		ctx context.Context, serviceID string, ttl int64, safePoint uint64,
-	) (uint64, error) {
-		t.Fatal("must not update")
-		return 0, nil
-	}
-	changefeedID1 := model.DefaultChangeFeedID("test-changefeed1")
-	tester.MustUpdate(
-		fmt.Sprintf("%s/changefeed/info/%s",
-			etcd.DefaultClusterAndNamespacePrefix,
-			changefeedID1.ID),
-		[]byte(`{"config":{},"state":"failed"}`))
-	tester.MustApplyPatches()
-	state.Changefeeds[changefeedID1].PatchStatus(
-		func(status *model.ChangeFeedStatus) (*model.ChangeFeedStatus, bool, error) {
-			return &model.ChangeFeedStatus{CheckpointTs: 2}, true, nil
-		})
-	tester.MustApplyPatches()
-	err = o.updateGCSafepoint(ctx, state)
-	require.Nil(t, err)
-
-	// switch the state of changefeed to normal, it must update GC safepoint to
-	// 1 (checkpoint Ts of changefeed-test1).
-	ch := make(chan struct{}, 1)
-	mockPDClient.UpdateServiceGCSafePointFunc = func(
-		ctx context.Context, serviceID string, ttl int64, safePoint uint64,
-	) (uint64, error) {
-		// Owner will do a snapshot read at (checkpointTs - 1) from TiKV,
-		// set GC safepoint to (checkpointTs - 1)
-		require.Equal(t, safePoint, uint64(1))
-		require.Equal(t, serviceID, etcd.GcServiceIDForTest())
-		ch <- struct{}{}
-		return 0, nil
-	}
-	state.Changefeeds[changefeedID1].PatchInfo(
-		func(info *model.ChangeFeedInfo) (*model.ChangeFeedInfo, bool, error) {
-			info.State = model.StateNormal
-			return info, true, nil
-		})
-	tester.MustApplyPatches()
-	err = o.updateGCSafepoint(ctx, state)
-	require.Nil(t, err)
-	select {
-	case <-time.After(5 * time.Second):
-		t.Fatal("timeout")
-	case <-ch:
-	}
-
-	// add another changefeed, it must update GC safepoint.
-	changefeedID2 := model.DefaultChangeFeedID("test-changefeed2")
-	tester.MustUpdate(
-		fmt.Sprintf("%s/changefeed/info/%s",
-			etcd.DefaultClusterAndNamespacePrefix,
-			changefeedID2.ID),
-		[]byte(`{"config":{},"state":"normal"}`))
-	tester.MustApplyPatches()
-	state.Changefeeds[changefeedID1].PatchStatus(
-		func(status *model.ChangeFeedStatus) (*model.ChangeFeedStatus, bool, error) {
-			return &model.ChangeFeedStatus{CheckpointTs: 20}, true, nil
-		})
-	state.Changefeeds[changefeedID2].PatchStatus(
-		func(status *model.ChangeFeedStatus) (*model.ChangeFeedStatus, bool, error) {
-			return &model.ChangeFeedStatus{CheckpointTs: 30}, true, nil
-		})
-	tester.MustApplyPatches()
-	mockPDClient.UpdateServiceGCSafePointFunc = func(
-		ctx context.Context, serviceID string, ttl int64, safePoint uint64,
-	) (uint64, error) {
-		// Owner will do a snapshot read at (checkpointTs - 1) from TiKV,
-		// set GC safepoint to (checkpointTs - 1)
-		require.Equal(t, safePoint, uint64(19))
-		require.Equal(t, serviceID, etcd.GcServiceIDForTest())
-		ch <- struct{}{}
-		return 0, nil
-	}
-	err = o.updateGCSafepoint(ctx, state)
-	require.Nil(t, err)
-	select {
-	case <-time.After(5 * time.Second):
-		t.Fatal("timeout")
-	case <-ch:
-	}
 }
 
 // make sure handleJobs works well even if there is two different
@@ -681,44 +565,6 @@ WorkLoop:
 	require.NotNil(t, infos[cf1])
 	require.NotNil(t, infos[cf2])
 	require.Nil(t, infos[cf3])
-}
-
-func TestCalculateGCSafepointTs(t *testing.T) {
-	state := orchestrator.NewGlobalState(etcd.DefaultCDCClusterID)
-	expectMinTsMap := make(map[uint64]uint64)
-	expectForceUpdateMap := make(map[uint64]interface{})
-	o := ownerImpl{changefeeds: make(map[model.ChangeFeedID]*changefeed)}
-
-	for i := 0; i < 100; i++ {
-		cfID := model.DefaultChangeFeedID(fmt.Sprintf("testChangefeed-%d", i))
-		upstreamID := uint64(i / 10)
-		cfInfo := &model.ChangeFeedInfo{UpstreamID: upstreamID, State: model.StateNormal}
-		cfStatus := &model.ChangeFeedStatus{CheckpointTs: uint64(i)}
-		changefeed := &orchestrator.ChangefeedReactorState{
-			ID:     cfID,
-			Info:   cfInfo,
-			Status: cfStatus,
-		}
-		state.Changefeeds[cfID] = changefeed
-
-		// expectMinTsMap will be like map[upstreamID]{0, 10, 20, ..., 90}
-		if i%10 == 0 {
-			expectMinTsMap[upstreamID] = uint64(i)
-		}
-
-		// If a changefeed does not exist in ownerImpl.changefeeds,
-		// forceUpdate should be true.
-		if upstreamID%2 == 0 {
-			expectForceUpdateMap[upstreamID] = nil
-		} else {
-			o.changefeeds[cfID] = nil
-		}
-	}
-
-	minCheckpoinTsMap, forceUpdateMap := o.calculateGCSafepoint(state)
-
-	require.Equal(t, expectMinTsMap, minCheckpoinTsMap)
-	require.Equal(t, expectForceUpdateMap, forceUpdateMap)
 }
 
 // AsyncStop should cleanup jobs and reject.
