+++ cdc/owner/server_manager.go
@@ -1,0 +1,242 @@
+// Copyright 2023 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package owner
+
+import (
+	"context"
+	"time"
+
+	"github.com/pingcap/errors"
+	"github.com/pingcap/failpoint"
+	"github.com/pingcap/log"
+	"github.com/pingcap/tiflow/cdc/model"
+	"github.com/pingcap/tiflow/pkg/config"
+	"github.com/pingcap/tiflow/pkg/orchestrator"
+	"github.com/pingcap/tiflow/pkg/upstream"
+	"github.com/pingcap/tiflow/pkg/version"
+	"go.uber.org/zap"
+	"golang.org/x/time/rate"
+)
+
+// ServerManager is a manager to schedule changefeeds
+type ServerManager interface {
+	orchestrator.Reactor
+}
+
+type serverManager struct {
+	changefeeds     map[model.ChangeFeedID]interface{}
+	upstreamManager *upstream.Manager
+
+	// logLimiter controls cluster version check log output rate
+	logLimiter   *rate.Limiter
+	lastTickTime time.Time
+	// bootstrapped specifies whether the owner has been initialized.
+	// This will only be done when the owner starts the first Tick.
+	// NOTICE: Do not use it in a method other than tick unexpectedly,
+	//         as it is not a thread-safe value.
+	bootstrapped bool
+
+	cfg *config.SchedulerConfig
+}
+
+// NewServerManager creates a new ServerManager
+func NewServerManager(
+	upstreamManager *upstream.Manager,
+	cfg *config.SchedulerConfig,
+) ServerManager {
+	return &serverManager{
+		upstreamManager: upstreamManager,
+		changefeeds:     make(map[model.ChangeFeedID]interface{}),
+		lastTickTime:    time.Now(),
+		logLimiter:      rate.NewLimiter(versionInconsistentLogRate, versionInconsistentLogRate),
+		cfg:             cfg,
+	}
+}
+
+// Tick implements the Reactor interface
+func (o *serverManager) Tick(stdCtx context.Context, rawState orchestrator.ReactorState) (nextState orchestrator.ReactorState, err error) {
+	failpoint.Inject("owner-run-with-error", func() {
+		failpoint.Return(nil, errors.New("owner run with injected error"))
+	})
+	failpoint.Inject("sleep-in-owner-tick", nil)
+	state := rawState.(*orchestrator.GlobalReactorState)
+	// At the first Tick, we need to do a bootstrap operation.
+	// Fix incompatible or incorrect meta information.
+	if !o.bootstrapped {
+		o.Bootstrap(state)
+		o.bootstrapped = true
+		return state, nil
+	}
+
+	if !o.clusterVersionConsistent(state.Captures) {
+		return state, nil
+	}
+	// Owner should update GC safepoint before initializing changefeed, so
+	// changefeed can remove its "ticdc-creating" service GC safepoint during
+	// initializing.
+	//
+	// See more gc doc.
+	if err = o.updateGCSafepoint(stdCtx, state); err != nil {
+		return nil, errors.Trace(err)
+	}
+
+	// Tick all changefeeds.
+	// ctx := stdCtx.(cdcContext.Context)
+	for _, changefeed := range state.Changefeeds {
+		o.changefeeds[changefeed.ID] = struct{}{}
+	}
+
+	return state, nil
+}
+
+// Bootstrap checks if the state contains incompatible or incorrect information and tries to fix it.
+func (o *serverManager) Bootstrap(state *orchestrator.GlobalReactorState) {
+	log.Info("Start bootstrapping")
+	fixChangefeedInfos(state)
+}
+
+// fixChangefeedInfos attempts to fix incompatible or incorrect meta information in changefeed state.
+func fixChangefeedInfos(state *orchestrator.GlobalReactorState) {
+	for _, changefeedState := range state.Changefeeds {
+		if changefeedState != nil {
+			changefeedState.PatchInfo(func(info *model.ChangeFeedInfo) (*model.ChangeFeedInfo, bool, error) {
+				if info == nil {
+					return nil, false, nil
+				}
+				info.FixIncompatible()
+				return info, true, nil
+			})
+		}
+	}
+}
+
+func (o *serverManager) clusterVersionConsistent(captures map[model.CaptureID]*model.CaptureInfo) bool {
+	versions := make(map[string]struct{}, len(captures))
+	for _, capture := range captures {
+		versions[capture.Version] = struct{}{}
+	}
+
+	if err := version.CheckTiCDCVersion(versions); err != nil {
+		if o.logLimiter.Allow() {
+			log.Warn("TiCDC cluster versions not allowed",
+				zap.String("ownerVer", version.ReleaseVersion),
+				zap.Any("captures", captures), zap.Error(err))
+		}
+		return false
+	}
+	return true
+}
+
+func (o *serverManager) updateGCSafepoint(
+	ctx context.Context, state *orchestrator.GlobalReactorState,
+) error {
+	minChekpoinTsMap, forceUpdateMap := o.calculateGCSafepoint(state)
+
+	for upstreamID, minCheckpointTs := range minChekpoinTsMap {
+		up, ok := o.upstreamManager.Get(upstreamID)
+		if !ok {
+			upstreamInfo := state.Upstreams[upstreamID]
+			up = o.upstreamManager.AddUpstream(upstreamInfo)
+		}
+		if !up.IsNormal() {
+			log.Warn("upstream is not ready, skip",
+				zap.Uint64("id", up.ID),
+				zap.Strings("pd", up.PdEndpoints))
+			continue
+		}
+
+		// When the changefeed starts up, CDC will do a snapshot read at
+		// (checkpointTs - 1) from TiKV, so (checkpointTs - 1) should be an upper
+		// bound for the GC safepoint.
+		gcSafepointUpperBound := minCheckpointTs - 1
+
+		var forceUpdate bool
+		if _, exist := forceUpdateMap[upstreamID]; exist {
+			forceUpdate = true
+		}
+
+		err := up.GCManager.TryUpdateGCSafePoint(ctx, gcSafepointUpperBound, forceUpdate)
+		if err != nil {
+			return errors.Trace(err)
+		}
+	}
+	return nil
+}
+
+// ignoreFailedChangeFeedWhenGC checks if a failed changefeed should be ignored
+// when calculating the gc safepoint of the associated upstream.
+func (o *serverManager) ignoreFailedChangeFeedWhenGC(
+	state *orchestrator.ChangefeedReactorState,
+) bool {
+	upID := state.Info.UpstreamID
+	us, exist := o.upstreamManager.Get(upID)
+	if !exist {
+		log.Warn("upstream not found", zap.Uint64("ID", upID))
+		return false
+	}
+	// in case the changefeed failed right after it is created
+	// and the status is not initialized yet.
+	ts := state.Info.StartTs
+	if state.Status != nil {
+		ts = state.Status.CheckpointTs
+	}
+	return us.GCManager.IgnoreFailedChangeFeed(ts)
+}
+
+// calculateGCSafepoint calculates GCSafepoint for different upstream.
+// Note: we need to maintain a TiCDC service GC safepoint for each upstream TiDB cluster
+// to prevent upstream TiDB GC from removing data that is still needed by TiCDC.
+// GcSafepoint is the minimum checkpointTs of all changefeeds that replicating a same upstream TiDB cluster.
+func (o *serverManager) calculateGCSafepoint(state *orchestrator.GlobalReactorState) (
+	map[uint64]uint64, map[uint64]interface{},
+) {
+	minCheckpointTsMap := make(map[uint64]uint64)
+	forceUpdateMap := make(map[uint64]interface{})
+
+	for changefeedID, changefeedState := range state.Changefeeds {
+		if changefeedState.Info == nil {
+			continue
+		}
+
+		switch changefeedState.Info.State {
+		case model.StateNormal, model.StateStopped, model.StateError:
+		case model.StateFailed:
+			if o.ignoreFailedChangeFeedWhenGC(changefeedState) {
+				continue
+			}
+		default:
+			continue
+		}
+
+		checkpointTs := changefeedState.Info.GetCheckpointTs(changefeedState.Status)
+		upstreamID := changefeedState.Info.UpstreamID
+
+		if _, exist := minCheckpointTsMap[upstreamID]; !exist {
+			minCheckpointTsMap[upstreamID] = checkpointTs
+		}
+
+		minCpts := minCheckpointTsMap[upstreamID]
+
+		if minCpts > checkpointTs {
+			minCpts = checkpointTs
+			minCheckpointTsMap[upstreamID] = minCpts
+		}
+		// Force update when adding a new changefeed.
+		_, exist := o.changefeeds[changefeedID]
+		if !exist {
+			forceUpdateMap[upstreamID] = nil
+		}
+	}
+	return minCheckpointTsMap, forceUpdateMap
+}
