+++ pkg/orchestrator/etcd_worker.go
@@ -133,7 +133,7 @@ func (worker *EtcdWorker) initMetrics() {
 // And the specified etcd session is nil-safety.
 func (worker *EtcdWorker) Run(ctx context.Context, session *concurrency.Session, timerInterval time.Duration, role string) error {
 	defer worker.cleanUp()
-	worker.isOwner = role == pkgutil.RoleOwner.String()
+	worker.isOwner = role == pkgutil.RoleServerManager.String()
 	// migrate data here
 	err := worker.checkAndMigrateMetaData(ctx, role)
 	if err != nil {
@@ -588,7 +588,7 @@ func (worker *EtcdWorker) checkAndMigrateMetaData(
 		return nil
 	}
 
-	if role != pkgutil.RoleOwner.String() {
+	if role != pkgutil.RoleServerManager.String() {
 		err := worker.migrator.WaitMetaVersionMatched(ctx)
 		return errors.Trace(err)
 	}
